requirements.txt
Flask
Flask-Babel
Flask-Oauthlib
Flask-Markdown

Flask-Migrate
Flask-SQLAlchemy

Pillow
pdfkit
sentry-sdk

blinker
